+++ server/base/ClientHeavyLoad.h
@@ -34,20 +34,20 @@ public slots:
     virtual void askLogin(const quint8 &query_id,const QString &login,const QByteArray &hash);
     virtual void askLoginBot(const quint8 &query_id);
     //check each element of the datapack, determine if need be removed, updated, add as new file all the missing file
-    void datapackList(const quint8 &query_id,const QStringList &files,const QList<quint32> &timestamps);
+    void datapackList(const quint8 &query_id, const QStringList &files, const QList<quint64> &timestamps);
     void dbQuery(const QString &queryText);
     void askedRandomNumber();
     //normal slots
     void askIfIsReadyToStop();
 private:
     // ------------------------------
-    bool sendFile(const QString &fileName,const QByteArray &content,const quint32 &mtime);
+    bool sendFile(const QString &fileName, const QByteArray &content, const quint64 &mtime);
     QString SQL_text_quote(QString text);
     // ------------------------------
     Player_internal_informations *player_informations;
     /** \brief send file if new or need be updated
      * \return return false if need be removed */
-    bool sendFileIfNeeded(const QString &filePath,const QString &fileName,const quint32 &mtime,const bool &checkMtime=true);
+    bool sendFileIfNeeded(const QString &filePath, const QString &fileName, const quint64 &mtime, const bool &checkMtime=true);
     void listDatapack(const QString &suffix,const QStringList &files);
     bool loadTheRawUTF8String();
     void loginIsRight(const quint8 &query_id,quint32 id,Map* map,const /*COORD_TYPE*/ quint8 &x,const /*COORD_TYPE*/ quint8 &y,const Orientation &orientation);
